Move onboard store to stores/
